issue-132: rename OlsDatatable component and remove word-break in Onto ID column
